requirements.txt
Django==3.1.3
celery==5.0.2
python-decouple
django-cors-headers==3.5.0
djangorestframework==3.12.2
whitenoise==5.2.0
psycopg2==2.8.6
Jinja2==3.1.2
openpyxl==3.1.2
pandas==2.0.2
cx-Oracle==8.0.1
requests==2.25.0
xlwt==1.3.0
pusher==3.0.0
Pillow
gunicorn==20.0.4
redis==3.5.0
django-redis==4.12.1
